+++ test/kpi-calculation.spec.ts
@@ -86,15 +86,15 @@ describe("[SDK] KPICalculationClient", () => {
         ControlSystemEvents: [
             {
                 _time: "2017-01-01T01:29:57.000Z",
-                type: "NORMAL_STOP",
+                type: KPICalculationModels.ControlSystemEvent.TypeEnum.NORMALSTOP,
             },
             {
                 _time: "2017-01-01T01:30:00.000Z",
-                type: "SHUTDOWN",
+                type: KPICalculationModels.ControlSystemEvent.TypeEnum.SHUTDOWN,
             },
             {
                 _time: "2017-01-01T04:30:01.000Z",
-                type: "SHUTDOWN",
+                type: KPICalculationModels.ControlSystemEvent.TypeEnum.SHUTDOWN,
             },
         ],
         calendar: {
@@ -223,19 +223,51 @@ describe("[SDK] KPICalculationClient", () => {
         const from = new Date();
         from.setDate(from.getDate() - 2);
         // console.log(data);
-        const result = await kpiCalculationClient.CaclulateKpiStates(
-            stateCalculationData as KPICalculationModels.RequestParametersBundle,
+        const result = await kpiCalculationClient.CaclulateKpiStates(stateCalculationData, {
+            from: new Date(data[0]["_time"]),
+            to: new Date(data[data.length - 1]["_time"]),
+            variableName: "kpiStatus",
+            initialState: "RSH",
+            defaultState: "FOH",
+            threshold: 1.1,
+            shutdownCorrelationThreshold: 5000,
+        });
+
+        result.should.not.be.undefined;
+    });
+
+    it("should calculate KPI states using direct method", async () => {
+        sdk.should.not.be.undefined;
+        kpiCalculationClient.should.not.be.undefined;
+
+        if (process.env.CI) {
+            return;
+        }
+
+        const today = new Date();
+        const yesterday = new Date(today);
+        yesterday.setDate(yesterday.getDate() - 1);
+        const from = new Date();
+        from.setDate(from.getDate() - 2);
+        const result = await kpiCalculationClient.CalculateKpiStatesDirect(
             {
-                from: new Date(data[0]["_time"]),
-                to: new Date(data[data.length - 1]["_time"]),
-                variableName: "kpiStatus",
-                initialState: "RSH",
-                defaultState: "FOH",
+                calendar: {},
+                ControlSystemEvents: [],
+            },
+            {
+                from: yesterday,
+                to: today,
+                variableName: "Temperature",
+                assetId: "4e268313c91d4dbebade263322f55369",
+                aspectName: "Environment",
+                initialState: "POH",
+                defaultState: "SH",
                 threshold: 1.1,
-                shutdownCorrelationThreshold: 5000,
+                shutdownCorrelationThreshold: 3000,
             }
         );
 
         result.should.not.be.undefined;
+        result.indications?.length.should.be.greaterThan(0);
     });
 });