+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements 3.10.0
 
+- SDK: fixed bug in EventManagementSDK GetEventTypes method (superflous history parameter)
 - unit tests: changed passkey for unit tests to work with passkey set up in environment (#207)
 
 ## 3.9.0 - (Pewter Vienna) - October 2020