+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 3.10.0 - (Seafoam Vienna)
 
-### Improvements 3.10.0
+### Bugfixes and Improvements 3.10.0
 
+- CLI: added support for direct calculation of KPI states to CLI mc kpi-calculation command
 - SDK: feat: add support for direct interaction to TrendPrediction Client (#204) [#hacktoberfest, coding4funrocks]
 - SDK: added support for direct interaction to KPI Calculation Client (#203) [#hacktoberfest, coding4funrocks]
 - SDK: fixed bug in EventManagementSDK GetEventTypes method (superflous history parameter)